Get MySQL up and running
 - Define root username/password in my55.cnf file
 - Run shell scripts as vagrant user
 - Make directories if they don't exist and change ownership to user: vagrant
 - Handy script to start/stop all components

+++ scripts/setup_mysql_ldconf_file.sh
@@ -4,7 +4,7 @@ source ./base.sh
 
 MY_LD_FILE=/etc/ld.so.conf.d/mysql.conf
 
-rm $MY_LD_FILE
+[ -e "$MY_LD_FILE" ] && rm $MY_LD_FILE
 
 echo "$PKG_DIR_BASE/mysqlclient/lib/mysql" > $MY_LD_FILE
 echo "$PKG_DIR_BASE/mysql/lib/mysql" >> $MY_LD_FILE
